% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate}
\alias{.simulate}
\title{Simulate seat allocations}
\usage{
.simulate(seats, shares, sigmas, N, threshold, ...)
}
\arguments{
\item{seats}{An integer with the total number of seats}

\item{shares}{A vector of vote shares for each party.}

\item{sigmas}{A vector of vote share uncertainty for ech party.}

\item{N}{An integer with the number of simulations.}

\item{threshold}{The electoral threshold (between 0 and 1)}

\item{...}{Additional arguments (currently not used)}
}
\value{
A matrix with the simulated seat distribution
}
\description{
Simulate seat allocations
}
