% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares2seats.R
\name{shares2seats}
\alias{shares2seats}
\title{Allocate seats to parties according to their shares}
\usage{
shares2seats(seats, shares, threshold = 0.03)
}
\arguments{
\item{seats}{An integer with the total number of seats}

\item{shares}{A vector with the vote share (between 0 and 1) for each party}

\item{threshold}{The electoral threshold (between 0 and 1)}
}
\value{
A vector of length \code{length(shares)} with the number of
seats allocated to each party
}
\description{
Allocate seats to parties according to their shares
}
\examples{
\dontrun{
res2021 <- c(25.0, 20.4, 17.9, 7.8, 7.8, 6.3, 6.1, 4.0)/100
shares2seats(85, res2021)
}
}
